scrapers/Tracxn_API/requirements.txt
2captcha-python==1.5.1
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dotenv==0.9.9
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mysql-connector-python==9.4.0
playwright==1.55.0
playwright-recaptcha==0.5.1
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
pyee==13.0.0
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
SpeechRecognition==3.10.4
starlette==0.48.0
temp-mail-so==1.0.2
tenacity==8.3.0
TwoCaptcha==0.0.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0

# AI/ML dependencies for similarity search
numpy>=1.26.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
torch>=2.0.0
